Avoid using puts for debug-messages
`puts` does bypass the logger the following that bypass output-formatting of gems like airbrussh. Therefore, I changed it to `debug` which invokes the logger and adds the right level. Now the output can be formatted according to the wishes of the user.

In order to avoid the time-penalty of collecting the data, the wrapping conditional still makes sense. If the user does not want to see debug-messages, they probably should not be collected in the first place. (See #42)